SCB-2094 Decouple rbac and etcd client
