+++ datasource/account.go
@@ -19,11 +19,18 @@ package datasource
 
 import (
 	"context"
+	"errors"
 	"github.com/apache/servicecomb-service-center/pkg/rbacframe"
+)
+
+var (
+	ErrDuplicated = errors.New("account is duplicated")
+	ErrCanNotEdit = errors.New("account can not be edited")
 )
 
 // AccountManager contains the RBAC CRUD
 type AccountManager interface {
+	CreateAccount(ctx context.Context, a *rbacframe.Account) error
 	AccountExist(ctx context.Context, key string) (bool, error)
 	GetAccount(ctx context.Context, key string) (*rbacframe.Account, error)
 	ListAccount(ctx context.Context, key string) ([]*rbacframe.Account, int64, error)
