+++ datasource/etcd/account.go
@@ -18,11 +18,58 @@ package etcd
 import (
 	"context"
 	"encoding/json"
+	"fmt"
+	"github.com/apache/servicecomb-service-center/datasource"
 	"github.com/apache/servicecomb-service-center/datasource/etcd/client"
 	"github.com/apache/servicecomb-service-center/datasource/etcd/kv"
+	"github.com/apache/servicecomb-service-center/pkg/etcdsync"
 	"github.com/apache/servicecomb-service-center/pkg/log"
 	"github.com/apache/servicecomb-service-center/pkg/rbacframe"
+	"github.com/apache/servicecomb-service-center/pkg/util"
+	stringutil "github.com/go-chassis/foundation/string"
+	"golang.org/x/crypto/bcrypt"
 )
+
+func (ds *DataSource) CreateAccount(ctx context.Context, a *rbacframe.Account) error {
+	lock, err := etcdsync.Lock("/account-creating/"+a.Name, -1, false)
+	if err != nil {
+		return fmt.Errorf("account %s is creating", a.Name)
+	}
+	defer func() {
+		err := lock.Unlock()
+		if err != nil {
+			log.Errorf(err, "can not release account lock")
+		}
+	}()
+	key := kv.GenerateETCDAccountKey(a.Name)
+	exist, err := datasource.Instance().AccountExist(ctx, a.Name)
+	if err != nil {
+		log.Errorf(err, "can not save account info")
+		return err
+	}
+	if exist {
+		return datasource.ErrDuplicated
+	}
+	hash, err := bcrypt.GenerateFromPassword([]byte(a.Password), 14)
+	if err != nil {
+		log.Errorf(err, "pwd hash failed")
+		return err
+	}
+	a.Password = stringutil.Bytes2str(hash)
+	a.ID = util.GenerateUUID()
+	value, err := json.Marshal(a)
+	if err != nil {
+		log.Errorf(err, "account info is invalid")
+		return err
+	}
+	err = client.PutBytes(ctx, key, value)
+	if err != nil {
+		log.Errorf(err, "can not save account info")
+		return err
+	}
+	log.Info("create new account: " + a.ID)
+	return nil
+}
 
 func (ds *DataSource) AccountExist(ctx context.Context, key string) (bool, error) {
 	resp, err := client.Instance().Do(ctx, client.GET,
